backend/requirements.txt
django
djangorestframework
django-cors-headers
python-dotenv
openai
google-generativeai
pillow